requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.2.1
